+++ LeadershipCollective/Repositories/IConsultantMessageRepository.cs
@@ -8,6 +8,8 @@ namespace LeadershipCollective.Repositories
         List<ConsultantRecMessage> GetAll();
         ConsultantRecMessage GetById(int id);
         void Add(ConsultantRecMessage consultantRecMessage);
+        void Update(ConsultantRecMessage consultantRecMessage);
+
         void Delete(int id); 
     }
 }
